I've checked in the following:
- Changes in the value hierarchy to incorporate a strict equality virtual method.
- Created a static push demultiplexer and a test dataflow.  The demultiplexer receives a vector of ValueRefs representing the demultiplexing keys in order. A tuple whose first field matches the i-th value in the demultiplexing key vector is forwarded to the i-th output.  The Demux element pushes back only if all of its outputs have pushed back.

The tests/demultiplex test creates a source data flow that creates time tuples every second, skipping seconds ending in 0, and prepending a 0 or a 1 according to whether the number of seconds is even or odd.

The demultiplexer switches output dataflows according to this 0 or 1.

Through no fault of my own, the demux key vector can hold disparate types.  So, for instance, output 0 may be selected when the first field is "alb", while output 1 is selected when the first field is the i32 number 9.  We could enforce that all demux keys are of the same type, but I don't quite see the point right now.