As per my earlier email, we figured out a way to deal with aggregates on the transient results of a join (e.g., to obtain the best finger during a lookup).  The solution is akin to building a computing iterator.  You define a startup PEL script, a per-found-tuple script (updating a common state for the aggregate), and a cleanup script.  For example, to compute a min, the startup script would put the maximum value (as initial min) and the comparator in the stack, the per-tuple-script would perform any filtering on tuples (e.g., of those fingers that overshoot the destination), and update the min in the stack, and the cleanup script would emit the result.
I've created a PelScan element that takes a table and an index and performs on it the above functionality, when given a scan tuple in its push input.  Its output is pull.  The final tuple (emitted during cleanup) is tagged with END_OF_SCAN.

Along the way, I also created the ()id, (]id, [)id, and []id PEL operators, we now have look-back operators for the PEL stack ("<uint> peek" means duplicate the uint-th stack element at the top of the stack), and a conditional execution stop ("<bool> ifstop" means "stop execution if the bool is true).  Ifstop is good for filtering out tuples during scanning. In the process, the PEL stack became a "staque" (i.e., a stack explicitly implemented on a deque, to allow random peeks into the chewy middle).  Also, not Pel_VM requires an explicit stack reset() before execute() is called, if you want the stack to be clean.  Otherwise, the existing stack in the VM is retained.  This way, we can carry over state from program execution to program execution during scans.

The tests/chord.C file contains a use case showing how these aggregates can be coded in PEL and put together to perform this functionality.  It essentially implements rule L2 (only the join with the node singleton tuple is assumed to have happened already).  The general idea is that as long as a rule requires any number of joins of a stream with a bunch of tables, we can do any combination of the three types of aggregates that Boon wrote about last night, by joining the stream with each table in succession.  We don't seem to have any rules that join more than one streams quite yet; there's never more than one stream that is not materialized.

To test rule L2, I also went ahead and built a simple functor source element.  This element takes a functor object with a single TupleRef 0-arg functor, which it runs whenever it is pulled.  Now we can pretty much parametrize the pull source to do arbitrary things, such as issuing random lookups (such a functor is implemented in chord.C).