Now table destructor kills primary index and flushes out queue if there
Turned all static search entries private to their functions to avoid
potential overruns in the future, esp. since they need not be reentrant
right now.

Fixed bug in insert, whereby the iterator of an entry to be replaced was
not advanced after the old tuple was removed. As a result, a subsequent
insertion of the replacement started out with an iterator pointing to a
new deleted entry.

Removed extraneous parameter from removeTuple (since it takes an element
position, it doesn't need also the tuple itself)

Iterators now clear their queues before destroying themselves