jol: (Mostly) replace the usage of Hashtable with HashMap. HashMap is preferable in modern Java code, because it is a proper Java 1.2 collections member, it isn't synchronized (and hence is faster), and as a bonus it also allows null keys.
Along the way, change various places (including the Table abstract
class) to reference the Map interface, not the particular concrete
implementation we're using (Hashtable or HashMap). This is just per
good OO design principles (information hiding). Making Hashtable
part of the Table API was particularly infelicitous.

In passing, fix what were likely 2 bugs, in Predicate#operator()
and Rule#localize(). The previous coding used Hashtable#contains(),
which actually checks if the given object is a ''value'' of the
hash table, not a key -- the code in both places actually meant
to do Map#containsKey(), as far as I can tell.