jol trunk: Land recent network race condition fixes on trunk.
Avoid a race condition in the network layer. We need to call
evaluate() after installing the OLG file for the network layer implementation,
but before we accept any connections. If this is not done, we risk accepting
connections before the associated OLG state has been created. Also add
an evaluate() for network.olg, to be safe.

Also, fix a typo in the UDP layer ("tcp" -> "udp").